add authentication

+++ app/hello.py
@@ -3,7 +3,9 @@ from flask_restful import Resource, Api
 from flask_sqlalchemy import SQLAlchemy
 from flask import Blueprint, request, render_template, flash, g, session, redirect, url_for
 from flask.ext.login import LoginManager
- 
+
+from flask import Flask,session, request, flash, url_for, redirect, render_template, abort ,g
+from flask_login import login_user , logout_user , current_user , login_required
 
 import os, datetime
 
@@ -11,23 +13,21 @@ _basedir = os.path.abspath(os.path.dirname(__file__))
 print datetime.datetime.now()
 
 
-try:
-    username = os.environ['user']
-    password = os.environ['pass']
-except KeyError as e:
-    username = 'bogie'
-    password = 'bogie'
 
 
 app = Flask(__name__)
 api = Api(app)
 
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + os.path.join(_basedir, 'bogie.db')
+app.config['SECRET_KEY'] = 'this is my secret key anf its very difficult to guess(but why..!)'
+
 
 db = SQLAlchemy(app)
 
 login_manager = LoginManager()
 login_manager.init_app(app)
+
+
 
 
 class Logz(db.Model):
@@ -60,106 +60,171 @@ class User(db.Model):
     password = db.Column('password' , db.String(10))
     # email = db.Column('email',db.String(50),unique=True , index=True)
     registered_on = db.Column('registered_on' , db.DateTime)
- 
-    def __init__(self , username ,password , email):
+    authenticated = db.Column(db.Boolean, default=False)
+
+    def __init__(self , username ,password):
         self.username = username
         self.password = password
-        self.email = email
-        self.registered_on = datetime.utcnow()
- 
-    def is_authenticated(self):
-        return True
+        # self.email = email
+        self.registered_on = datetime.datetime.now()
  
     def is_active(self):
+        """True, as all users are active."""
         return True
- 
+
+    def is_authenticated(self):
+        """Return True if the user is authenticated."""
+        return self.authenticated
+
     def is_anonymous(self):
+        """False, as anonymous users aren't supported."""
         return False
- 
     def get_id(self):
-        return unicode(self.id)
+        return self.username
  
+
     def __repr__(self):
-        return '<User %r>' % (self.username)
+        return '<User %r %r>' % (self.username,self.password)
 
 
+@login_manager.user_loader
+def user_loader(user_id):
+    return User.query.get(user_id)
 
+        
+
+
+try:
+    username = os.environ['user']
+    password = os.environ['pass']
+except KeyError as e:
+    username = 'bogie'
+    password = 'bogie'
+db.drop_all()                                                                                                                               
+db.create_all()
+userz = User(username,password)
+# dummy = logz('dev', 'dummy',"2.0",datetime.datetime.now(),"dummySource")
+db.session.add(userz)
+db.session.commit()
+
+
+ 
+@app.route('/login',methods=['GET','POST'])
+def login():
+    if request.method == 'GET':
+        return render_template('login.html')
+    qusername = request.form['username']
+    qpassword = request.form['password']
+    registered_user = User.query.filter_by(username=qusername,password=qpassword).first()
+    print registered_user
+    if registered_user is None:
+        return redirect(url_for('login'))
+    login_user(registered_user)
+    session['logged_in'] = True
+    return render_template('index.html')
 
 # later on
 # db.init_app(app)
 
 
-# db.drop_all()                                                                                                                               
-# db.create_all()
-# dummy = logz('dev', 'dummy',"2.0",datetime.datetime.now(),"dummySource")
-# db.session.add(dummy)
-# db.session.commit()
+
+
+
 
 @app.route('/', methods=['GET'])
+# @login_required
 def index():
-    return render_template('index.html')
+    if  not session.get('logged_in'):
+        return redirect(url_for('login'))
+    else:
+        return render_template('index.html')
+
 
 
 @app.route('/current', methods=['GET'])
 def current():
-    query = db.session.query(Logz.components.distinct()).all()
-    systms = db.session.query(Logz.systems.distinct()).all()
-    print systms
+    if  not session.get('logged_in'):
+        return redirect(url_for('login'))
+    else:
+        query = db.session.query(Logz.components.distinct()).all()
+        systms = db.session.query(Logz.systems.distinct()).all()
+        print systms
 
-    log = []
-    for x in query:
-        for y in systms:
-            qs = Logz.query.filter_by(components=x[0], systems=y[0]).order_by(Logz.timestamp.desc(), Logz.systems).first()
-            if qs is not None:
-                log.append(qs)
-            # print x[0]
-            # log.append(Logz.query.filter_by(components=x[0]).order_by(Logz.timestamp.desc()).first())
-    return render_template('current.html', list_of_items=log)
+        log = []
+        for x in query:
+            for y in systms:
+                qs = Logz.query.filter_by(components=x[0], systems=y[0]).order_by(Logz.timestamp.desc(), Logz.systems).first()
+                if qs is not None:
+                    log.append(qs)
+                # print x[0]
+                # log.append(Logz.query.filter_by(components=x[0]).order_by(Logz.timestamp.desc()).first())
+        return render_template('current.html', list_of_items=log)
 
 
 @app.route('/current/components/<string:components>', methods=['GET'])
 def current_comp(components):
-    query = db.session.query(Logz.systems.distinct()).filter_by(components=components).all()
-    log = []
-    for x in query:
-        qs = Logz.query.filter_by(systems=x[0], components=components).order_by(Logz.timestamp.desc()).first()
-        if qs is not None:
-            log.append(qs)
-        
-    return render_template('current.html', list_of_items=log)
+    if  not session.get('logged_in'):
+        return redirect(url_for('login'))
+    else:
+        query = db.session.query(Logz.systems.distinct()).filter_by(components=components).all()
+        log = []
+        for x in query:
+            qs = Logz.query.filter_by(systems=x[0], components=components).order_by(Logz.timestamp.desc()).first()
+            if qs is not None:
+                log.append(qs)
+            
+        return render_template('current.html', list_of_items=log)
 
 
 @app.route('/current/systems/<string:systems>', methods=['GET'])
 def current_sys(systems):
-    query = db.session.query(Logz.components.distinct()).filter_by(systems=systems).all()
-    log = []
-    for x in query:
-        # print x[0]
-        qs =Logz.query.filter_by(components=x[0], systems=systems).order_by(Logz.timestamp.desc()).first()
-        if qs is not None:
-            log.append(qs)
+    if  not session.get('logged_in'):
+        return redirect(url_for('login'))
+    else:
+        query = db.session.query(Logz.components.distinct()).filter_by(systems=systems).all()
+        log = []
+        for x in query:
+            # print x[0]
+            qs =Logz.query.filter_by(components=x[0], systems=systems).order_by(Logz.timestamp.desc()).first()
+            if qs is not None:
+                log.append(qs)
 
-    return render_template('current.html', list_of_items=log)
+        return render_template('current.html', list_of_items=log)
 
 
 # @app.route('/get',methods = ['GET'])
 @app.route('/logs', methods=['GET'])
 def get_logs():
-    log = Logz.query.order_by(Logz.timestamp.desc()).all()
-    return render_template('hello.html', list_of_items=log)
+    if  not session.get('logged_in'):
+        return redirect(url_for('login'))
+    else:
+        log = Logz.query.order_by(Logz.timestamp.desc()).all()
+        return render_template('hello.html', list_of_items=log)
 
 
 @app.route('/logs/<string:components>', methods=['GET'])
 def get__component_logs(components):
-    list_of_items = Logz.query.filter_by(components=components).order_by(Logz.timestamp.desc())
-    return render_template('hello.html', list_of_items=list_of_items)
+    if  not session.get('logged_in'):
+        return redirect(url_for('login'))
+    else:
+        list_of_items = Logz.query.filter_by(components=components).order_by(Logz.timestamp.desc())
+        return render_template('hello.html', list_of_items=list_of_items)
 
 
 @app.route('/syslogs/<string:systems>', methods=['GET'])
 def get__system_logs(systems):
-    list_of_items = Logz.query.filter_by(systems=systems).order_by(Logz.timestamp.desc())
-    return render_template('hello.html', list_of_items=list_of_items)
+    if  not session.get('logged_in'):
+        return redirect(url_for('login'))
+    else:
+        list_of_items = Logz.query.filter_by(systems=systems).order_by(Logz.timestamp.desc())
+        return render_template('hello.html', list_of_items=list_of_items)
 
+
+@app.route('/logout')
+def logout():
+    logout_user()
+    session['logged_in'] = False
+    return redirect(url_for('index')) 
 
 
 @app.route('/test', methods=['GET'])
@@ -170,32 +235,53 @@ def test():
 
 class TodoSimple(Resource):
     def put(self, ):
+        susername = request.form['user']
+        spass = request.form['pass']
         systems = request.form['systems']
         components = request.form['components']
         version = request.form['version']
         timestamp = datetime.datetime.utcnow()
         source = request.form['source']
-        u = Logz(systems, components, version, timestamp, source)
-        db.session.add(u)
-        db.session.commit()
-        return {'done': True}
+        if username == 'bogie' and password == 'bogie' :
+            spass = 'bogie'
+            susername = 'bogie' 
+        
+        if username==susername and password==spass:
+            u = Logz(systems, components, version, timestamp, source)
+            db.session.add(u)
+            db.session.commit()
+            return {'done': True}
+        else:
+            return {'Error': "Unauthorised Access"}
+
         # return "{'systems': %s, 'components': %s, 'version': %s, 'time': %s, 'source': %s}"% systems,components,version,time,source
 
     def post(self):
+        susername = request.form['user']
+        spass = request.form['pass']
         components = request.form['components']
         version = request.form['version']
         systems = request.form['systems']
         timestamp = request.form['timestamp']
         source = request.form['source']
         nextz = request.form['next']
-        u = Logz.query.filter_by(systems=systems, components=components, version=version, timestamp=timestamp,
-                                 source=source).first()
-        if u is not None:
-            db.session.delete(u)
-            db.session.commit()
-            # return {'status':True}
-            return redirect('/' + nextz)
-        return {'status': False}
+
+        if username == 'bogie'and password == 'bogie' :
+            spass = 'bogie'
+            susername = 'bogie' 
+        
+        if username==susername and password == spass:
+            u = Logz.query.filter_by(systems=systems, components=components, version=version, timestamp=timestamp,
+                                     source=source).first()
+            if u is not None:
+                db.session.delete(u)
+                db.session.commit()
+                # return {'status':True}
+                return redirect('/' + nextz)
+            return {'status': False}
+        else:
+            return {'Error': "Unauthorised Access"}
+
 
 
 # @app.route('/delete', methods=['POST'])
